requirements.txt
anyio==4.8.0
black==24.10.0
boolean.py==4.0
CacheControl==0.14.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cyclonedx-python-lib==7.6.2
defusedxml==0.7.1
filelock==3.16.1
flake8==7.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
license-expression==30.4.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
mypy-extensions==1.0.0
nyct-gtfs==2.0.0
packageurl-python==0.16.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
platformdirs==4.3.6
protobuf==4.25.3
psutil==6.1.1
py-serializable==1.1.2
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
